refactor: Extract validation classes into their own location.
