+++ tools/serverpod_cli/lib/src/analyzer/models/validation/restrictions.dart
@@ -1156,81 +1156,3 @@ class Restrictions {
     return classDefinition.tableName != null;
   }
 }
-
-class StringValueRestriction {
-  List<SourceSpanSeverityException> validate(
-    String parentNodeName,
-    dynamic value,
-    SourceSpan? span,
-  ) {
-    if (value is! String) {
-      return [
-        SourceSpanSeverityException(
-          'The property must be a String.',
-          span,
-        )
-      ];
-    }
-
-    return [];
-  }
-}
-
-class EnumValueRestriction<T extends Enum> {
-  List<T> enums;
-
-  EnumValueRestriction({
-    required this.enums,
-  });
-
-  List<SourceSpanSeverityException> validate(
-    String parentNodeName,
-    dynamic enumValue,
-    SourceSpan? span,
-  ) {
-    var options = enums.map((v) => v.name);
-
-    var errors = <SourceSpanSeverityException>[
-      SourceSpanSeverityException(
-        '"$enumValue" is not a valid property. Valid properties are $options.',
-        span,
-      )
-    ];
-
-    if (enumValue is! String) return errors;
-
-    var isEnumValue = enums.any(
-      (e) => e.name.toLowerCase() == enumValue.toLowerCase(),
-    );
-
-    if (!isEnumValue) return errors;
-
-    return [];
-  }
-}
-
-class BooleanValueRestriction {
-  List<SourceSpanSeverityException> validate(
-    String parentNodeName,
-    dynamic value,
-    SourceSpan? span,
-  ) {
-    if (value is bool) return [];
-
-    var errors = [
-      SourceSpanSeverityException(
-        'The value must be a boolean.',
-        span,
-      )
-    ];
-
-    if (value is! String) return errors;
-
-    var boolValue = value.toLowerCase();
-    if (!(boolValue == 'true' || boolValue == 'false')) {
-      return errors;
-    }
-
-    return [];
-  }
-}
