+++ tools/serverpod_cli/lib/src/analyzer/models/yaml_definitions/exception_yaml_definition.dart
@@ -1,5 +1,6 @@
 import 'package:serverpod_cli/src/analyzer/models/validation/keywords.dart';
 import 'package:serverpod_cli/src/analyzer/models/validation/restrictions.dart';
+import 'package:serverpod_cli/src/analyzer/models/validation/restrictions/base.dart';
 import 'package:serverpod_cli/src/analyzer/models/validation/validate_node.dart';
 
 class ExceptionYamlDefinition {
